backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
psycopg2-binary>=2.9.9
python-dotenv==1.0.1
websockets==12.0
apscheduler==3.10.4
requests==2.32.3
twilio==9.2.3
osmnx==2.0.1
networkx==3.3
shapely>=2.0
pyproj>=3.6
bcrypt>=4.0.1
google-auth>=2.29.0
geopandas>=0.14.0